Various improvements
I fixed the PreventExitant trigger once and for all!
Tags are now set with a separate (REQUIRED) input field on newpage. As such, I removed the tagging prompt, and instead a message will appear on pages lacking them.
Atabang is no more! To center text use `->TEXT<-`

'potential' pages will appear on books properly now, and the grimoire will prompt for an overview instead of tags.
I made various changes, and variables will now test for none using `is` and `is not` instead of `==` and `!=`

Grimoire_Config is now considerably more useful, allowing you to change the port, enable and disable markdown, potential pages, and prompts

the header's <title> element now reads Grimoire, and I threw in an empty database.