Timeline
Added a timeline!
I removed Age as a prompt and replaced it with birthdate.
Birthdate will add a point to the timeline. (A link to the timeline will be added to the main book page at a later date)
Adding "Event" to a page will give you a prompt for date, which adds that event and it's overview to the timeline.
I made a few other changes, as well as updating the database with a 'Dates' table to make the timeline possible.
I changed how the Linked trigger worked so that after every insert it will clean up links to existing pages.